requirements.txt
streamlit
pandas
plotly
python-dateutil
dateparser
google-generativeai==0.8.5
python-dotenv
matplotlib